requirements.txt
yt-dlp
tqdm
deepgram-sdk
openai
llama-index